# expression: implement vectorized evaluation for `builtinSubDateAndDurationSig`