@shihongzhi 2019-12-09
Duplicated with pr #13691